Modification AlphaTreeNodeDescriptorSyntax pour ne recalculer la valeur qu'une fois qu'il y a un certain nombres de nouveaux pixels + ajout de la plus petite liste à la plus grande.

+++ Student-Project-2015/src/students/desanlis_pierrel/descriptor/AlphaTreeNodeDescriptorContourRatio.java
@@ -57,7 +57,7 @@ public class AlphaTreeNodeDescriptorContourRatio extends AlphaTreeNodeFilterDesc
 		listPixel.add(coord);
 		nouveauPts++;
 
-		if (nouveauPts/listPixel.size() > ratioMaj){			
+		if ((double)nouveauPts/listPixel.size() > ratioMaj){			
 			value = calcValue(); 
 			if (value < min)
 				min = value;
@@ -92,7 +92,7 @@ public class AlphaTreeNodeDescriptorContourRatio extends AlphaTreeNodeFilterDesc
 			nouveauPts+= desc.listPixel.size();
 		}
 
-		if (nouveauPts/listPixel.size() > ratioMaj)
+		if ((double)nouveauPts/listPixel.size() > ratioMaj)
 		{
 			value = calcValue(); 
 			if (value < min)
